周一晚

+++ database/migrations/2018_01_20_093432_mission_templates.php
@@ -32,7 +32,7 @@ class MissionTemplates extends Migration
         
         $table->float('sustain');
         
-        $table->float('arithmetic');
+        $table->integer('arithmetic');
         
         /**
          * Foreignkeys section
